wip(be): define "Round" & "Game" models
In order to make the app scalable, we need to include "Round" & "Game" models.
In each "Round" we will have a clue, user's answer, correct answer, etc as its fields.
The fields of the Game model are not clear in my mind at the moment, perhaps "n" number of rounds can be a one game.